Start operations in a background thread, whether they are asynchronous or not
Per the docs

> Operation queues use the Dispatch framework to initiate the execution of their operations. As a result, queues always invoke operations on a separate thread, regardless of whether the operation is synchronous or asynchronous.

While we're not using dispatch in this case, starting them in a background thread better fits the spec and fixes client code that depends on this behaviour.

We no longer call `finish` from the background thread - if the operation is synchronous, it will call `finish` or equivalent before returning. If it is asynchronous, it will call `finish` or equivalent at a later time