Implement Copy On Write pattern for NSCalendar
NSCalendar now holds a single pointer to a GSCalendarData. GSCalendarData is a read-only view into an ICU calendar with a given:
- identifier
- locale
- timeZone
- firstWeekday
- minimumDaysInFirstWeek

GSCalendarData retains ownership of the underlying ICU calendar in order to call ucal_close once it's deallocated.

This allows `-[NSCalendar copyWithZone:]` to be implemented by a simple pointer copy of GSCalendarData, without needing to reopen nor clone the existing ICU UCalendar.
Because GSCalendarData is read-only, we also avoid needing to lock/unlock the calendar to perform date calculations, since we can on-demand create clones as necessary

A clone will be made on-demand if we attempt to modify the NSCalendar properties